requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
spotipy==2.23.0
python-dotenv==1.0.0
Pillow==10.1.0



